+++ src/models/Book/BookImage.js
@@ -1,20 +1,20 @@
-const sequelize = require('../config/database');
+const sequelize = require('../../config/database');
 const { DataTypes } = require('sequelize');
 
-const EditorialCollection = sequelize.define(
-  'editorial_collection',
+const BookImage = sequelize.define(
+  'book_image',
   {
     id: {
       type: DataTypes.UUID,
       primaryKey: true,
       defaultValue: DataTypes.UUIDV4,
     },
-    name: {
-      type: DataTypes.STRING,
+    image: {
+      type: DataTypes.BLOB,
       allowNull: false,
     },
   },
   { timestamps: false, freezeTableName: true }
 );
 
-module.exports = EditorialCollection;
+module.exports = BookImage;
