+++ src/models/Sale/PaymentMethod.js
@@ -1,20 +1,21 @@
-const sequelize = require('../config/database');
+const sequelize = require('../../config/database');
+
 const { DataTypes } = require('sequelize');
 
-const ReviewLike = sequelize.define(
-  'review_like',
+const PaymentMethod = sequelize.define(
+  'payment_method',
   {
     id: {
       type: DataTypes.UUID,
       primaryKey: true,
       defaultValue: DataTypes.UUIDV4,
     },
-    review_status: {
-      type: DataTypes.BOOLEAN,
+    name: {
+      type: DataTypes.STRING,
       allowNull: false,
     },
   },
   { timestamps: false, freezeTableName: true }
 );
 
-module.exports = ReviewLike;
+module.exports = PaymentMethod;
