+++ src/models/Shelves/BookShelfCategory.js
@@ -1,4 +1,4 @@
-const sequelize = require('../config/database');
+const sequelize = require('../../config/database');
 
 const { DataTypes } = require('sequelize');
