+++ src/models/Shelves/BookShelves.js
@@ -1,4 +1,4 @@
-const sequelize = require('../config/database');
+const sequelize = require('../../config/database');
 const { DataTypes } = require('sequelize');
 
 const BookShelves = sequelize.define(
